refactor method injection
