+++ imports/api/diExample/methods.js
@@ -1,6 +1,3 @@
-import '/imports/startup/server/diContainer';
-
-
 Meteor.injectedMethods({
   'diExample.searchClient'(query) {
     const { hmisClient, logger } = this.context;
@@ -8,4 +5,13 @@ Meteor.injectedMethods({
     logger.debug('got results', results);
     return results;
   },
+  'diExample.test'(...args) {
+    // logger.debug('got args', args);
+    const { sentryScope, logger } = this.context;
+    sentryScope.addBreadcrumb({ message: 'this is a sample message' });
+    console.log('aaa', Object.keys(this.context), this.context.endpointName);
+    throw new Error('abcd');
+    return [args, logger];
+  },
+
 });
